# docs/openapi/rekurn.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rekurn API",
    "version": "0.1.0",
    "description": "Public API for Rekurn repositories, refs, objects, deployments, audit logs, and passkey authentication."
  },
  "servers": [{ "url": "https://api.rekurn.com/api/v1" }],
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
...
      "Repo": { "type": "object", "properties": { "name": { "type": "string" }, "description": { "type": ["string", "null"] }, "visibility": { "enum": ["public", "private"] }, "defaultBranch": { "type": "string" }, "deployHooks": { "type": "object", "additionalProperties": { "type": "string", "format": "uri" } }, "createdAt": { "type": "string", "format": "date-time" } }, "required": ["name", "visibility", "defaultBranch", "createdAt"] },
...
    }
  },
  "paths": {
...
    "/repos": {
      "get": { "summary": "List repositories", "responses": { "200": { "description": "Repositories", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Repo" } } } } } } },
      "post": { "summary": "Create repository", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "description": { "type": "string" }, "visibility": { "enum": ["public", "private"] }, "defaultBranch": { "type": "string" } }, "required": ["name"] } } } }, "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Repo" } } } } } }
    },
    "/repos/{ownerId}/{name}": {
      "parameters": [{ "name": "ownerId", "in": "path", "required": true, "schema": { "type": "string" } }, { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": { "summary": "Get repository", "responses": { "200": { "description": "Repository", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Repo" } } } } } },
      "delete": { "summary": "Delete repository", "responses": { "200": { "description": "Deleted" } } }
    },
...
  }
}